move generic implementation of sqrt_ratio to SqrtRatio
